backend/requirements.txt
fastapi==0.109.2
uvicorn==0.27.1
pydantic==2.6.1
openai==1.12.0
python-dotenv==1.0.0
httpx==0.26.0
nvidia-nim-sdk==0.1.0
cloudflare==2.19.0
together==0.2.5
python-jose==3.3.0
passlib==1.7.4
bcrypt==4.1.2
pytest==8.0.0
pytest-asyncio==0.23.5 